more cleanup of bindings
