+++ bridge/ffi/src/utils.rs
@@ -1,0 +1,70 @@
+use crate::bindings::*;
+use dsnp_graph_config::{ConnectionType, DsnpVersion, PrivacyType};
+use std::collections::HashMap;
+
+// Function to convert C-compatible `SchemaConfig` to a Rust `SchemaConfig`
+pub fn schema_config_from_ffi(schema_config: &SchemaConfig) -> dsnp_graph_config::SchemaConfig {
+	let dsnp_version = match schema_config.dsnp_version {
+		DsnpVersion::Version1_0 => dsnp_graph_config::DsnpVersion::Version1_0,
+	};
+
+	let connection_type = match &schema_config.connection_type {
+		ConnectionType::Follow(privacy_type) => {
+			let privacy = match privacy_type {
+				PrivacyType::Public => dsnp_graph_config::PrivacyType::Public,
+				PrivacyType::Private => dsnp_graph_config::PrivacyType::Private,
+			};
+			dsnp_graph_config::ConnectionType::Follow(privacy)
+		},
+		ConnectionType::Friendship(privacy_type) => {
+			let privacy = match privacy_type {
+				PrivacyType::Public => dsnp_graph_config::PrivacyType::Public,
+				PrivacyType::Private => dsnp_graph_config::PrivacyType::Private,
+			};
+			dsnp_graph_config::ConnectionType::Friendship(privacy)
+		},
+	};
+
+	dsnp_graph_config::SchemaConfig { dsnp_version, connection_type }
+}
+
+// Function to convert C-compatible `Config` to a Rust `Config`
+pub fn config_from_ffi(config: &Config) -> dsnp_graph_config::Config {
+	let schema_map_slice =
+		unsafe { std::slice::from_raw_parts(config.schema_map, config.schema_map_len) };
+	let mut schema_map = HashMap::new();
+	for schema_config_tuple in schema_map_slice {
+		let id = schema_config_tuple.schema_id;
+		let schema_config = &schema_config_tuple.schema_config;
+		schema_map.insert(id, schema_config_from_ffi(schema_config));
+	}
+
+	let dsnp_versions_slice =
+		unsafe { std::slice::from_raw_parts(config.dsnp_versions, config.dsnp_versions_len) };
+	let mut dsnp_versions = Vec::new();
+	for version in dsnp_versions_slice {
+		let rust_version = match version {
+			DsnpVersion::Version1_0 => dsnp_graph_config::DsnpVersion::Version1_0,
+		};
+		dsnp_versions.push(rust_version);
+	}
+	dsnp_graph_config::Config {
+		sdk_max_users_graph_size: config.sdk_max_users_graph_size,
+		sdk_max_stale_friendship_days: config.sdk_max_stale_friendship_days,
+		max_graph_page_size_bytes: config.max_graph_page_size_bytes,
+		max_page_id: config.max_page_id,
+		max_key_page_size_bytes: config.max_key_page_size_bytes,
+		schema_map,
+		dsnp_versions,
+	}
+}
+
+// Function to convert C-compatible `SchemaConfig` to a Rust `SchemaConfig`
+pub fn environment_from_ffi(environment: &Environment) -> dsnp_graph_config::Environment {
+	let config = config_from_ffi(&environment.config);
+	match environment.environment_type {
+		EnvironmentType::Mainnet => dsnp_graph_config::Environment::Mainnet,
+		EnvironmentType::Rococo => dsnp_graph_config::Environment::Rococo,
+		EnvironmentType::Dev => dsnp_graph_config::Environment::Dev(config),
+	}
+}
